requirements.txt
anchorpy-core==0.2.0
annotated-types==0.7.0
anyio==3.7.1
apischema==0.16.6
attrs==24.2.0
base58==2.1.1
based58==0.1.1
borsh-construct==0.1.0
cachetools==4.2.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
construct==2.10.68
construct-typing==0.5.6
distro==1.9.0
fastapi==0.115.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.6.1
jsonalias==0.1.1
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
more-itertools==8.14.0
openai==1.52.2
OSlash==0.6.3
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyheck==0.1.5
PyNaCl==1.5.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.20.0
sniffio==1.3.1
solana==0.21.0
solders==0.21.0
starlette==0.41.0
sumtypes==0.1a6
tabulate==0.9.0
toml==0.10.2
toolz==0.11.2
tqdm==4.66.5
types-cachetools==4.2.10
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==10.4
gunicorn==20.1.0
